migrate test_cassandra_mod to pytest
